Исправления наставника. На защиту.
